+++ examples/remoter/_request/main.go
@@ -19,7 +19,7 @@ const (
 			"host":"hidden-lake-service=remoter",
 			"path":"/exec",
 			"head":{
-				"Hl-Remoter-Password": "DpxJFjAlrs4HOWga0wk14mZqQSBo9DxK"
+				"Password": "DpxJFjAlrs4HOWga0wk14mZqQSBo9DxK"
 			},
 			"body":"%s"
 		}